+++ app/__init__.py
@@ -1,5 +1,5 @@
 """Root package"""
-from .dependancies import *
+from .classifier import *
 from .models import *
 from .routers import *
 from .config import *
